refactor: extract camera implementation to in-house component
