+++ src/main/home/send/SendScan/index.tsx
@@ -7,21 +7,18 @@ import React, {
   useState,
 } from 'react';
 import {
-  Dimensions,
   StyleSheet,
   TouchableOpacity,
   View,
   Image,
   ScrollView,
 } from 'react-native';
-import { RNHoleView } from 'react-native-hole-view';
 import BottomSheet, { BottomSheetView } from '@gorhom/bottom-sheet';
 import { BarcodeFormat, useScanBarcodes } from 'vision-camera-code-scanner';
-import { useCameraDevices } from 'react-native-vision-camera';
-import { Camera } from 'react-native-vision-camera';
 import { quais } from 'quais';
 import { styledColors } from 'src/shared/styles';
 import {
+  QuaiPayCamera,
   QuaiPayContent,
   QuaiPayListItem,
   QuaiPayLoader,
@@ -30,7 +27,7 @@ import {
 } from 'src/shared/components';
 import { t } from 'i18next';
 import { useUsername, useWallet } from 'src/shared/hooks';
-import { useIsFocused, useNavigation } from '@react-navigation/native';
+import { useNavigation } from '@react-navigation/native';
 import { RootStackNavigationProps } from 'src/shared/navigation';
 import { useTheme } from 'src/shared/context/themeContext';
 import { useContacts } from 'src/shared/hooks/useContacts';
@@ -38,12 +35,7 @@ import { Contact, Theme } from 'src/shared/types';
 import DownChevronBlack from 'src/shared/assets/downChevronBlack.svg';
 import DownChevronWhite from 'src/shared/assets/downChevronWhite.svg';
 
-const windowWidth = Dimensions.get('window').width;
-const squareSize = windowWidth * 0.65;
-const squarePaddingRight = (windowWidth - squareSize) / 2;
-
 function SendScanScreen() {
-  const isFocused = useIsFocused();
   const navigation = useNavigation<RootStackNavigationProps<'Main'>>();
   // const wallet = useWallet();
   const sender = useUsername();
@@ -69,7 +61,6 @@ function SendScanScreen() {
     sheetRef.current?.close();
   }, []);
 
-  const [hasPermission, setHasPermission] = useState(false);
   const [bottomSheetUp, setBottomSheetUp] = useState(false);
   const [searchText, setSearchText] = useState('');
   const [filteredContacts, setFilteredContacts] = useState<Contact[]>([]);
@@ -87,21 +78,10 @@ function SendScanScreen() {
       );
     }
   }, [contacts, searchText]);
-  const devices = useCameraDevices();
-  const device = devices.back;
 
   const [frameProcessor, barcodes] = useScanBarcodes([BarcodeFormat.QR_CODE], {
     checkInverted: true,
   });
-  const [isCameraReady, setIsCameraReady] = React.useState(true);
-
-  useEffect(() => {
-    if (isFocused) {
-      setIsCameraReady(true);
-    } else {
-      setIsCameraReady(false);
-    }
-  }, [isFocused]);
 
   const styles = themedStyle(theme);
 
@@ -136,49 +116,13 @@ function SendScanScreen() {
   //   runOnJS(setBarcodes)(detectedBarcodes);
   // }, []);
 
-  useEffect(() => {
-    (async () => {
-      const status = await Camera.requestCameraPermission();
-      setHasPermission(status === 'authorized');
-    })();
-  }, []);
-
   if (!wallet) {
     return <QuaiPayLoader text={t('loading')} />;
   }
 
   return (
     <QuaiPayContent noNavButton hasBackgroundVariant>
-      <View style={styles.cameraContainer}>
-        {device != null && hasPermission && isCameraReady && (
-          <>
-            <Camera
-              style={StyleSheet.absoluteFill}
-              device={device}
-              isActive={true}
-              frameProcessor={frameProcessor}
-              frameProcessorFps={5}
-            />
-            <RNHoleView
-              style={[
-                styles.holeView,
-                {
-                  backgroundColor: styledColors.black,
-                },
-              ]}
-              holes={[
-                {
-                  x: squarePaddingRight,
-                  y: 120,
-                  width: squareSize,
-                  height: squareSize,
-                  borderRadius: 10,
-                },
-              ]}
-            />
-          </>
-        )}
-      </View>
+      <QuaiPayCamera frameProcessor={frameProcessor} />
       {contacts ? (
         <BottomSheet
           backgroundStyle={{
@@ -310,21 +254,7 @@ const themedStyle = (theme: Theme) =>
     container: {
       flex: 1,
     },
-    cameraContainer: {
-      flex: 1,
-      position: 'absolute',
-      top: 0,
-      left: 0,
-      right: 0,
-      bottom: 0,
-    },
     image: { width: 40, height: 40, borderRadius: 20 },
-    holeView: {
-      position: 'absolute',
-      width: '100%',
-      height: '100%',
-      opacity: 0.6,
-    },
     bottomSheetContainer: {
       paddingHorizontal: 24,
       flexDirection: 'row',
